refactor(web): move COE premiums to /coe/premiums with new landing page
