+++ apps/web/src/app/(main)/(explore)/coe/premiums/components/fun-facts-pqp-section.tsx
