+++ apps/web/src/app/(main)/(explore)/coe/premiums/components/key-insights.tsx
